LayoutTests/storage/indexeddb/deletedatabase-delayed-by-versionchange-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test the order when there are pending setVersion, delete and open calls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "deletedatabase-delayed-by-versionchange.html"
ver = 1
setVersionBlockedEventFired = false
versionChangeComplete = false
deleteDatabaseBlockedEventFired = false
deleteDatabaseComplete = false
indexedDB.open(dbname)

openOnSuccess():
h = event.target.result
indexedDB.open(dbname)

h2OpenOnSuccess():
h.setVersion(String(ver++))
indexedDB.deleteDatabase(dbname)

h2OnVersionChange():
old = 1
new = "1"

setVersionOnBlocked():
setVersionBlockedEventFired = true
indexedDB.open(dbname)
h2.close()

setVersionOnSuccess():

transactionOnComplete():
versionChangeComplete = true

onVersionChange():
old = "1"
new = ""

deleteDatabaseOnBlocked():
deleteDatabaseBlockedEventFired = true
h.close()

deleteDatabaseOnSuccess():
deleteDatabaseComplete = true

h3OpenOnSuccess():
PASS setVersionBlockedEventFired is true
PASS versionChangeComplete is true
PASS deleteDatabaseBlockedEventFired is true
PASS deleteDatabaseComplete is true
PASS successfullyParsed is true

TEST COMPLETE
